pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Etore-BeS <etorebraga@usp.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
sqlalchemy = "^2.0.29"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pytest-asyncio = "^0.23.6"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
pytest = "^8.1.1"
taskipy = "^1.12.2"
ruff = "^0.4.1"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warning"

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'uvicorn fast_zero.app:app --reload'
pre-test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

